+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 ### Compatibility
 * This release is compatible with:
   * Kotlin 1.6.10.
-  * Coroutines 1.5.2-native-mt.
+  * Coroutines 1.6.0-native-mt. Also compatible with Coroutines 1.6.0 but requires enabling of the new memory model and disabling of freezing, see https://github.com/realm/realm-kotlin#kotlin-memory-model-and-coroutine-compatibility for details on that.
   * AtomicFu 0.17.0.
 * Minimum Gradle version: 6.1.1.  
 * Minimum Android Gradle Plugin version: 4.0.0.